Implement SubscriptionInput base class
